requirements.txt
PyPDF2>=3.0.0
pdf2image>=1.16.3
Pillow>=10.0.0
python-dotenv>=1.0.0
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
pytesseract>=0.3.10
PyQt6>=6.4.0 